fix(studio): repair rebuilt timeline integration
