+++ .fallowrc.jsonc
@@ -88,8 +88,79 @@
     // the sync-hyperframes-codegen.yml workflow; complexity/dead-code
     // findings on this file are not actionable from this repo.
     "packages/cli/src/cloud/_gen/**",
+    // TEMP(studio-dnd): rebuilt Ular modules intentionally remain unwired until
+    // the follow-up Studio shell cutover lands. Keep their exact payload without
+    // treating the coexistence copies as production dead code or duplication.
+    "packages/studio/src/components/EditorShell.tsx",
+    "packages/studio/src/components/nle/AssetPreviewOverlay.tsx",
+    "packages/studio/src/components/nle/NLEContext.tsx",
+    "packages/studio/src/components/nle/PreviewOverlays.tsx",
+    "packages/studio/src/components/nle/PreviewPane.tsx",
+    "packages/studio/src/components/nle/TimelinePane.tsx",
+    "packages/studio/src/components/nle/useTimelineEditCallbacks.ts",
+    "packages/studio/src/components/sidebar/AssetCard.tsx",
+    "packages/studio/src/hooks/useTimelineEditingDrops.ts",
   ],
   "ignoreExports": [
+    // TEMP(studio-dnd): public helpers consumed by the remaining shell-cutover
+    // payloads after this rebuilt stack.
+    {
+      "file": "packages/studio/src/components/editor/domEditOverlayGeometry.ts",
+      "exports": ["overlayCornersCentroid", "elementCornerOverlayPoints", "orientedOverlayRect"],
+    },
+    {
+      "file": "packages/studio/src/components/editor/manualOffsetDrag.ts",
+      "exports": ["applyManualOffsetNudgeDraft", "applyManualOffsetNudgeCommit"],
+    },
+    {
+      "file": "packages/studio/src/hooks/timelineEditingGsap.ts",
+      "exports": ["patchIframeRootDuration", "shiftGsapPositionsBatch", "syncTimingEditPreview"],
+    },
+    {
+      "file": "packages/studio/src/hooks/timelineEditingHelpers.ts",
+      "exports": ["patchIframeRootDuration", "shiftGsapPositionsBatch", "syncTimingEditPreview"],
+    },
+    {
+      "file": "packages/studio/src/player/components/timelineGroupEditing.ts",
+      "exports": [
+        "buildTimelineGroupResizeMembers",
+        "resolveTimelineGroupResizeChanges",
+        "applyTimelineGroupResizePreview",
+      ],
+    },
+    {
+      "file": "packages/studio/src/player/components/timelineEditing.ts",
+      "exports": [
+        "selectTimelineElementsInMarquee",
+        "TimelineGroupResizeEdge",
+        "TimelineGroupTimingMember",
+      ],
+    },
+    {
+      "file": "packages/studio/src/player/lib/timelineElementHelpers.ts",
+      "exports": ["furthestClipEndFromDocument", "furthestClipEndFromSource"],
+    },
+    {
+      "file": "packages/studio/src/components/sidebar/AssetContextMenu.tsx",
+      "exports": ["DeleteConfirm"],
+    },
+    {
+      "file": "packages/studio/src/utils/timelineAssetDrop.ts",
+      "exports": [
+        "setCompositionDurationToContent",
+        "extendCompositionDurationIfNeeded",
+        "fitTimelineAssetGeometry",
+        "resolveTimelineAssetCompositionSize",
+      ],
+    },
+    {
+      "file": "packages/studio/src/components/sidebar/assetHelpers.ts",
+      "exports": ["truncateMiddle", "formatDuration"],
+    },
+    {
+      "file": "packages/studio/src/utils/studioHelpers.ts",
+      "exports": ["resolveDroppedAssetDimensions"],
+    },
     // drawElementService is the bottom of the fast-capture Graphite stack
     // (#1917): its consumers (frameCapture in #1919) land two PRs upstack, so
     // a per-PR audit diffing against the merge base sees these exports as
@@ -308,6 +379,15 @@
     // require intrusive middleware changes beyond this PR's scope.
     "minLines": 6,
     "ignore": [
+      // TEMP(studio-dnd): coexistence implementations and parallel test cases
+      // retained verbatim from the rebuilt Ular payload until shell cutover.
+      "packages/studio/src/player/components/ImageThumbnail.test.tsx",
+      "packages/studio/src/player/components/VideoThumbnail.test.tsx",
+      "packages/studio/src/hooks/timelineEditingHelpers.ts",
+      "packages/studio/src/hooks/useBlockHandlers.ts",
+      "packages/studio/src/utils/blockInstaller.ts",
+      "packages/studio/src/utils/timelineAssetDrop.ts",
+      "packages/studio/src/hooks/useAnimatedPropertyCommit.test.tsx",
       // sourcePatcher.ts: pre-existing internal clones between the inline-style
       // and attribute tag-patchers; only the PatchOperation type gained two
       // optional fields here, but the line shift makes fallow re-flag them.
@@ -433,6 +513,13 @@
     // complexity pre-dates the computed-timeline work. Exempted at file level
     // rather than refactored as scope creep.
     "ignore": [
+      // TEMP(studio-dnd): inherited/high-complexity helpers touched by the
+      // exact rebuilt payload; refactoring them is outside this recovery PR.
+      "packages/studio/src/player/lib/timelineIframeHelpers.ts",
+      "packages/studio/src/player/hooks/usePlaybackKeyboard.ts",
+      "packages/studio/src/player/lib/timelineElementHelpers.ts",
+      "packages/studio/src/components/editor/manualOffsetDrag.ts",
+      "packages/studio/src/components/editor/domEditOverlayGeometry.ts",
       // sourcePatcher.ts: resolveSourceFile / splitInlineStyleDeclarations /
       // patch*InTag pre-date this PR; only the PatchOperation type gained two
       // optional fields, but the line-shift fingerprint re-flags the inherited
